// frontend/web/src/api/types.gen/SetPipelineReq.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PipelineEdge } from "./PipelineEdge";
import type { PipelineKind } from "./PipelineKind";
import type { RouteDefinition } from "./RouteDefinition";

export type SetPipelineReq = { strategy_id: string, kind: PipelineKind, edges: Array<PipelineEdge>, route?: RouteDefinition, };